+++ config/bootstrap.php
@@ -20,19 +20,31 @@
  * @license     http://www.gnu.org/licenses/agpl.txt AGPL License
  * @link        http://git.novatlantis.it Nova Atlantis Ltd
  */
-
 use Cake\Core\Configure;
 
+//Sets the default Assets name
+if (!defined('ASSETS')) {
+    define('ASSETS', 'Assets');
+}
+
 //If `true`, assets will be used even if debugging is enabled
-if (!Configure::check('Assets.force')) {
-    Configure::write('Assets.force', false);
+if (!Configure::check(ASSETS . '.force')) {
+    Configure::write(ASSETS . '.force', false);
 }
 
 //Default assets directory
-if (!Configure::check('Assets.target')) {
-    Configure::write('Assets.target', TMP . 'assets');
+if (!Configure::check(ASSETS . '.target')) {
+    Configure::write(ASSETS . '.target', TMP . 'assets');
 }
 
-if (!is_writeable(Configure::read('Assets.target'))) {
-    trigger_error(sprintf('Directory %s not writeable', Configure::read('Assets.target')), E_USER_ERROR);
+//Checks for target directory
+$target = Configure::read(ASSETS . '.target');
+
+if (!file_exists($target)) {
+    //@codingStandardsIgnoreLine
+    @mkdir($target);
+}
+
+if (!is_writeable($target)) {
+    trigger_error(sprintf('Directory %s not writeable', $target), E_USER_ERROR);
 }